Refactor seperated Get and Post in different controllers
